Fix int32/int64 dtype mismatch in Windows CI
Integration test failed on Windows due to platform-specific integer types:

Problem:
- np.arange() returns int32 on Windows, int64 on Unix
- Validation requires explicit int64 for time columns
- Error: 'Unsupported dtype int32 for time column'

Solution:
- Added explicit dtype=np.int64 to np.arange() in time column tests
- Added explicit dtype=np.float64 to np.linspace() for consistency
- Ensures cross-platform compatibility

Tests affected:
- test_integer_time_column
- test_float_time_column

Result: All 163 integration tests pass on all platforms

+++ tests/integration/test_oop_functional_equivalence.py
@@ -406,11 +406,11 @@ class TestTimeColumnTypes:
 
     def test_integer_time_column(self):
         """Test equivalence with integer time column."""
-        # Create data with integer time
+        # Create data with integer time (explicit int64 for cross-platform compatibility)
         np.random.seed(111)
         data = pd.DataFrame(
             {
-                "day": np.arange(1, 51),
+                "day": np.arange(1, 51, dtype=np.int64),
                 "control": np.random.normal(100, 10, 50),
                 "test": np.random.normal(105, 11, 50),
             }
@@ -448,11 +448,11 @@ class TestTimeColumnTypes:
 
     def test_float_time_column(self):
         """Test equivalence with float time column."""
-        # Create data with float time
+        # Create data with float time (explicit float64 for cross-platform compatibility)
         np.random.seed(222)
         data = pd.DataFrame(
             {
-                "time": np.linspace(0.0, 10.0, 50),
+                "time": np.linspace(0.0, 10.0, 50, dtype=np.float64),
                 "control": np.random.normal(100, 10, 50),
                 "test": np.random.normal(105, 11, 50),
             }
